requirements.txt
attrs==20.2.0
certifi==2020.6.20
click==7.1.2
h11==0.9.0
httpcore==0.11.1
httpx==0.15.5
idna==2.10
iniconfig==1.0.1
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.0
python-dateutil==2.8.1
pytz==2020.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
toml==0.10.1
